Change binding identifier, add a class reference type to it
Not entirely sure that this is the correct way of doing this in TypeScript, but should make a little more sense when dealing with bindings.

Also, added dedicated binding methods for dealing with class references. This is not as elegant, but JavaScript is not able to determine whether a function is a class constructor or not (at least not in a safe way).